shows sd and mean. Uses Gaussian random for testing.

+++ app/src/main/java/com/cowman/turlough/androidcharts/StatsFragment.java
@@ -75,7 +75,7 @@ public class StatsFragment extends Fragment {
         public void onFragmentInteraction(Uri uri);
     }
 
-    public void add(int data){
+    public void add(double data){
         chart.add(data);
     }
